requirements.txt
langchain==0.1.0
langchain-google-genai==0.0.6
google-generativeai==0.3.2
faiss-cpu
gradio
pandas
datasets
python-dotenv